+++ packages/template-chakra-storefront/src/pages/cart/hooks/use-cart-operations.js
@@ -8,9 +8,9 @@ import {useState} from 'react'
 import {useIntl} from 'react-intl'
 import debounce from 'lodash/debounce'
 import {useShopperBasketsMutation} from '@salesforce/commerce-sdk-react'
-import useToast from '../use-toast'
-import {TOAST_MESSAGE_REMOVED_ITEM_FROM_CART} from '../../../config/constants'
-import {getUpdateBundleChildArray} from '../../utils/product-utils'
+import useToast from '../../../hooks/use-toast'
+import {TOAST_MESSAGE_REMOVED_ITEM_FROM_CART} from '../../../../config/constants'
+import {getUpdateBundleChildArray} from '../../../utils/product-utils'
 
 const DEBOUNCE_WAIT = 750
